Change FONT data type from int[][] to byte[][]

+++ src/chip8/Keyboard.java
@@ -22,7 +22,7 @@ package chip8;
  */
 
 public class Keyboard {
-    private static final int[][] FONT = {
+    private static final byte[][] FONT = {
             {0xF0, 0x90, 0x90, 0x90, 0xF0}, // 0
             {0x20, 0x60, 0x20, 0x20, 0x70}, // 1
             {0xF0, 0x10, 0xF0, 0x80, 0xF0}, // 2
